Make use of express middleware for error handling.
The current way of handling errors doesn't work well in async code.
This is a proposal how to make use of the way express does error handling.
Instead of throwing an error we pass it to the next function (third parameter to action).
The error will then be handled by an express middleware.

This is a proposal how this could be done for swagger-node-express 2.0.